pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "candlecrawl"
version = "0.1.0a1"
description = "Self-hostable crawling, scraping, mapping, search, and extraction infrastructure."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "CandleCrawl maintainers" }
]
keywords = ["crawling", "scraping", "search", "extraction", "fastapi", "playwright"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: FastAPI",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27,<0.28",
  "pydantic>=2.11,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
service = [
  "fastapi>=0.133,<0.134",
  "uvicorn>=0.35,<0.36",
  "pydantic-settings>=2.5,<3",
  "markdownify>=0.14,<0.15",
  "beautifulsoup4>=4.12,<5",
  "loguru>=0.7,<0.8",
  "slowapi>=0.1.9,<0.2",
  "prometheus-client>=0.20,<1",
  "redis>=5,<6",
  "rq>=1.16,<2",
  "tiktoken>=0.8,<0.9",
]
browser = ["playwright>=1.45,<2"]
pdf = ["pypdf>=6.7,<7"]
ocr = ["Pillow>=12,<13"]
dev = ["build>=1.2,<2", "twine>=5,<7", "ruff>=0.8,<1"]
test = ["pytest>=8.4,<9", "pytest-asyncio>=1,<2"]
all = [
  "fastapi>=0.133,<0.134",
  "uvicorn>=0.35,<0.36",
  "pydantic-settings>=2.5,<3",
  "markdownify>=0.14,<0.15",
  "beautifulsoup4>=4.12,<5",
  "loguru>=0.7,<0.8",
  "slowapi>=0.1.9,<0.2",
  "prometheus-client>=0.20,<1",
  "redis>=5,<6",
  "rq>=1.16,<2",
  "tiktoken>=0.8,<0.9",
  "playwright>=1.45,<2",
  "pypdf>=6.7,<7",
  "Pillow>=12,<13",
  "build>=1.2,<2",
  "twine>=5,<7",
  "ruff>=0.8,<1",
  "pytest>=8.4,<9",
  "pytest-asyncio>=1,<2",
]

[project.urls]
Homepage = "https://github.com/kmccleary3301/candlecrawl"
Repository = "https://github.com/kmccleary3301/candlecrawl"
Issues = "https://github.com/kmccleary3301/candlecrawl/issues"

[project.scripts]
candlecrawl = "candlecrawl.cli:main"

[tool.setuptools]
package-dir = { "candlecrawl" = "src/candlecrawl" }

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["candlecrawl*", "app*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
